+++ CHANGES.md
@@ -1,8 +1,9 @@
 # Release History
 
 ## 0.7.3
-- only initialize a http session when the first session value is written
+- only initialize a http session when it is actually used
 - make it possible to better configure the session
+- add session example
  
 ## 0.7.2
 - also support FormResponse for exceptions